lambda/send_alerts_compound/requirements.txt
requests==2.31.0
mysql-connector-python==8.0.32
sendgrid==6.10.0
pyyaml==6.0.1
